requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.8
django-autoslug==1.9.8
django-cors-headers==3.10.0
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.0
Faker==9.5.2
idna==3.3
inflection==0.5.1
ipython==7.28.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
progressbar2==3.55.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
traitlets==5.1.0
types-pytz==2021.3.1
types-PyYAML==6.0.1
typing_extensions==4.0.0
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
